backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioread==3.1.0
boto3==1.42.54
botocore==1.42.54
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
decorator==5.2.1
fastapi==0.129.0
ffmpeg-python==0.2.0
future==1.0.0
h11==0.16.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
jmespath==1.1.0
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
moviepy==2.2.1
msgpack==1.1.2
numba==0.64.0
numpy==2.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==11.3.0
platformdirs==4.9.2
pooch==1.9.0
proglog==0.1.12
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
